Make Error component, move Info to containers folder
