+++ src/Components/Error/Error.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Error = (props) => <h2 className="error-display">{props.message}</h2>
