+++ src/Components/Error/error.scss
@@ -1,0 +1,6 @@
+.error-display {
+  color: rgb(67, 93, 148);
+  font-size: 99%;
+  margin: 0;
+  padding: 0;
+}
