+++ src/Components/Header/_header.scss
@@ -1,7 +1,3 @@
-.header {
-  position: relative;
-}
-
 h1 {
   color: $title;
   margin: 0;
