+++ src/Containers/App/App.js
@@ -5,8 +5,9 @@ import { addPalettes } from "../../Actions";
 import { fetchData } from "../../Utils/API";
 import PropTypes from "prop-types";
 import { Loading } from "../../Components/Loading/Loading";
+import { Error } from "../../Components/Error/Error";
 import { Header } from "../../Components/Header/Header";
-import { Info } from "../../Components/Info/Info";
+import { Info } from "../Info/Info";
 import PalettePicker from "../PalettePicker/PalettePicker";
 
 export class App extends Component {
@@ -97,9 +98,7 @@ export class App extends Component {
             </div>
           </div>
           <div className="right-head">
-            {this.state.error && 
-              <h2 className="errer-display">{this.state.error}</h2>
-            }
+            {this.state.error && <Error message={this.state.error} />}
             {this.state.isLoading && <Loading />}
           </div>
         </div>
