+++ src/Containers/App/_app.scss
@@ -58,16 +58,8 @@ body {
   justify-content: space-around;
 }
 
-
 .main {
   grid-area: main;
-}
-
-.error-display {
-  color: rgb(67, 93, 148);
-  font-size: 99%;
-  margin: 0;
-  padding: 0;
 }
 
 .info-btn {
