+++ src/Containers/Info/info.scss
