+++ src/Style/index.scss
@@ -4,7 +4,8 @@
 @import '../Containers/App/app';
 @import '../Components/Header/header';
 @import '../Components/Loading/loading';
-@import '../Components/Info/info';
+@import '../Components/Error/error';
+@import '../Containers/Info/info';
 @import '../Components/InstructionBox/instructionBox';
 @import '../Containers/PalettePicker/palettePicker';
 @import '../Containers/Projects/projects';
